Hey lakifume!
I rebalanced a LOT in this one:
Alucard starts with a lot of HP
Enemies have lots of HP
You both deal high number damages
Equipments stats adjusted for this progression - even for INT (since it's the only basis for subweapon/magic damage)

First purpose of this balance is to prevent getting the STR > weaponATK to avoid the double damage later in the game
This also makes the progression of stats less steep - meaning the "challenge" is always there, but the difficulty slowly curves higher the further you are in the game.

(I'm still doing runs for this balancing so I might change something else...but if you want you can try it yourself and hope you like it! It's been great so far for me)